feat(gshell): K5c への追随 — CTRL+STOP はフォーカス窓宛、Quit 無視は 3 秒後に exec_kill (W-2、決裁 A1/A3)
- 型の追随: 束縛が i32 になったので multiapp.rs / lib.rs の往復キャストを整理 (挙動不変)。
- A1 (契約 T6): op_wait で abort_seen を見たとき、フォーカスが本人なら従来どおり break。
  別のアプリなら multiapp::redirect_abort() が「取り消し」と「フォーカスの ID の kill」を予約し、
  should_park の (a) (top-level にしか出来ない仕事) で譲らせ、top-level の drain_top_level() が
  (1) exec_abort_clear → (2) exec_kill の順で実行 (逆だと走っている本人が syscall 出口で巻き添え)。
  アプリが 1 本のときも (a) で top-level へ戻れる。
- A3: session::QUIT_GRACE_CYCLES = 300 (1 周 = wm_cycle → sys_halt = PIT 10ms → 3 秒の上界)。
  arm_quit (SWITCH_CUI / SHUTDOWN) で仕掛け、x3_cycle で減らし、0 で request_kill_all()。
  全アプリへ Quit → 待ち → 残りを kill → 全回収 → SessionAction。
D11 の規則は不変 ((a) の項は有限事象なので上界 30 は変わらない)。

ホスト試験: wm_tests.rs +5 本 (29 → 34)、mocks.rs に exec_abort_clear の差し替えと回数記録。
RED (空実装) で (b)(c) の 2 FAIL → GREEN 34/34。記録: k5b_gshell_tdd.md 追補 W-2 (回 8/9)。
ゲスト未検証 (CTRL+STOP の付け替え・Quit 無視の打ち切りは実機未実施)。
残り穴: gui_call を一切呼ばず CPU を握るアプリは WM の周期が来ないので猶予が進まない (カーネルの領分)。

+++ userland/gshell/src/lib.rs
@@ -258,10 +258,9 @@ fn run_program(st: &mut wm::GuiState, path: &[u8; 256]) -> i32 {
      * いる ID はまだ分かっていない (戻り値そのものなので) ので、譲り合いの表に
      * 「起動が進行中」の印を立てておく (`multiapp::begin_start` の注記)。 */
     multiapp::begin_start();
-    /* 生成物の Rust 束縛は `i32` を u32 として吐く (sdk/kapi_rust_gen.py の
-     * TYPE_MAP に "i32" が無い)。ABI は EAX の i32 のままなので、ここで戻す。
-     * 生成物は手で触らない ([ABI1])。 */
-    let rc = unsafe { (os32api::api().exec_start)(path.as_ptr()) as i32 };
+    /* KAPI v45 (K5c) で生成器が `i32` を吐くようになったので、u32 経由の
+     * 往復は要らない。生成物は手で触らない ([ABI1])。 */
+    let rc = unsafe { (os32api::api().exec_start)(path.as_ptr()) };
     multiapp::end_start(rc);
     /* 描画モードの復帰は **アプリが抜けたときだけ** (不具合 W-1、2026-09-11)。
      *
